Refactor Balance app: Convert to ES6 modules (analog to Simulator)
This commit modularizes the Balance app by splitting the monolithic
balance-app.js file into clean ES6 modules, following the same pattern
used in the Simulator app.

Changes:
- Created 7 ES6 modules with clear separation of concerns:
  * balance-config.js - Configuration and error classes
  * balance-utils.js - Pure utility functions
  * balance-storage.js - Persistence layer (LocalStorage, IndexedDB, File System API)
  * balance-reader.js - Input reading layer
  * balance-renderer.js - Output rendering layer (21 methods)
  * balance-binder.js - Event handling layer (17 methods + 2 helpers)
  * balance-main.js - Main orchestration and entry point

- Updated Balance.html to use ES6 module imports
- Added comprehensive BALANCE_MODULES_README.md documentation
- Removed old monolithic js/balance/balance-app.js

Benefits:
- Single responsibility principle - Each module has one clear purpose
- Dependency injection - Clean module boundaries
- Easier testing - Modules can be tested in isolation
- Better maintainability - Easy to find and modify code
- Improved code navigation - Organized by function type
- Modern tooling support - Hot module replacement, bundle splitting

Module structure mirrors Simulator app for consistency across the codebase.
All functionality from the original monolithic version preserved.

Version updated from v21.1 to v22.0 (ES6 Modules).
