+++ Balance.html
@@ -10,7 +10,7 @@
 
     <!-- External Scripts -->
     <script src="engine.js" defer></script>
-    <script src="js/balance/balance-app.js" defer></script>
+    <script type="module" src="balance-main.js"></script>
 </head>
 <body>
     <div class="main-layout">
